pyproject.toml
[tool.poetry]
name = "stockbinz"
version = "0.1.0"
description = ""
authors = ["deepak <info.deepakpatidar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^5.2.0"}
phonenumbers = "^8.12.31"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
